camera plugin constants now exported as $cordovaCameraConstants
The camera constants are used in the camera options. Without this
stubbed it was not possible to use the constants in code without errors.
The values used in the mock are directly copied from the source
of CordovaConstants.js that is added to a project when the
camera plugin is installed.