mock $cordovaCamera renamed to ngCordovaMocks.$cordovaCamera
Currently each cordova plugin wrapper and corresponding mock are named
the same. This means that both can't be included in a project at the
same time because there is no namespacing of components from different
modules. This forces a developer to have preprocessor step in their
build pipeline and doesn't work well in scenarios when a developer is
designing a responsive interface with new tools such as browsersync
http://www.browsersync.io/ .

The issue is also talked about here:

http://stackoverflow.com/questions/15342657/how-to-reference-values-with-the-same-name-from-different-modules

The suggestion given in the answer seems to be popular with many
developers:

> your best option is to prefix your values
> (and other providers) with a module name

This commit demonstrates what that suggestion looks like when
implemented. Subsequent commits will show how the
ngCordovaMocks.$cordovaCamera and other mocks can now easily be used
without name collisions.