Adds $cordovaCameraConstants and ngCordovaMocks.$cordovaCameraConstants
The `$cordovaCamera` plugin has a number of constants that are exposed
on `window` if cordova.js and cordova_plugins.js are included. When not
running on a device or in test and exception will be thrown if they are
accessed. CameraConstants.js is where constants are defined. The mock
contains a copy of the contents of CameraConstants.js.

ngCordovaMocks.$cordovaCameraConstants is injected in the test as
$cordovaCameraConstants. The same could be done in a `config()` of
a module if the application is running outside of a device

In the camera plugin test a `beforeEach()` is added the injects
`$window.Camera` with `ngCordovaMocks.$cordovaCameraConstants` in
case a test is added in the future that needs the value of one of
the constants