$cordovaScreenOrientation: can now build a boxed orientation by name
Given an orientation name from `$cordovaScreenOrientationConstants` it
is helpful to be able to get boxed object version of the name like you
already can for the current $window.orientation value. The
`$cordovaScreenOrientation` has been slightly refactored to allow for this
use case. It creates a new instance of boxed orientation each time the
`$cordovaScreenOrientation.get([orientationName])` is called

When the boxed orientation is a wrapper for `$window.orientation`
care is taken not assign the value `$window.orientation` to a local
variable. This ensures that the boxed orientation always gives the
correct answer if cached and reused.